requirements/local.txt
-r base.txt

WebOb==1.4
WebTest==2.0.15
beautifulsoup4==4.3.2
coverage==3.7.1
django-debug-toolbar==1.2.1
django-webtest==1.7.7
six==1.7.3
sqlparse==0.1.11
waitress==0.8.9
